Adding error code for not_found response and moving related code to separate file
