+++ backend/vaccination_api/cmd/digilocker_support_api/main.go
@@ -4,15 +4,8 @@ import (
 	"archive/zip"
 	"bytes"
 	"compress/flate"
-	"crypto/hmac"
-	"crypto/sha256"
-	"encoding/base64"
-	"encoding/hex"
 	"encoding/json"
-	"encoding/xml"
-	"fmt"
 	"github.com/divoc/api/config"
-	"github.com/divoc/api/pkg"
 	"github.com/divoc/api/pkg/auth"
 	"github.com/divoc/api/pkg/models"
 	kafkaService "github.com/divoc/api/pkg/services"
@@ -43,10 +36,6 @@ const InternalFailedEvent = "internal-failed"
 const ExternalSuccessEvent = "external-success"
 const ExternalFailedEvent = "external-failed"
 
-type ErrorResponse struct {
-	Status  string `json:"status"`
-	Message string `json:"message"`
-}
 
 type Certificate struct {
 	Context           []string `json:"@context"`
@@ -108,331 +97,6 @@ type Certificate struct {
 	} `json:"proof"`
 }
 
-type PullURIRequest struct {
-	XMLName    xml.Name `xml:"PullURIRequest"`
-	Text       string   `xml:",chardata"`
-	Ns2        string   `xml:"ns2,attr"`
-	Ver        string   `xml:"ver,attr"`
-	Ts         string   `xml:"ts,attr"`
-	Txn        string   `xml:"txn,attr"`
-	OrgId      string   `xml:"orgId,attr"`
-	Format     string   `xml:"format,attr"`
-	DocDetails struct {
-		Text         string `xml:",chardata"`
-		DocType      string `xml:"DocType"`
-		DigiLockerId string `xml:"DigiLockerId"`
-		UID          string `xml:"UID"`
-		FullName     string `xml:"FullName"`
-		DOB          string `xml:"DOB"`
-		Photo        string `xml:"Photo"`
-		TrackingId   string `xml:"tracking_id"`
-		Mobile       string `xml:"Mobile"`
-		UDF1         string `xml:"UDF1"`
-		UDF2         string `xml:"UDF2"`
-		UDF3         string `xml:"UDF3"`
-		UDFn         string `xml:"UDFn"`
-	} `xml:"DocDetails"`
-}
-
-type PullURIResponse struct {
-	XMLName        xml.Name `xml:"PullURIResponse"`
-	Text           string   `xml:",chardata"`
-	Ns2            string   `xml:"ns2,attr"`
-	ResponseStatus struct {
-		Text   string `xml:",chardata"`
-		Status string `xml:"Status,attr"`
-		Ts     string `xml:"ts,attr"`
-		Txn    string `xml:"txn,attr"`
-	} `xml:"ResponseStatus"`
-	DocDetails struct {
-		Text         string `xml:",chardata"`
-		DocType      string `xml:"DocType"`
-		DigiLockerId string `xml:"DigiLockerId"`
-		UID          string `xml:"UID"`
-		FullName     string `xml:"FullName"`
-		DOB          string `xml:"DOB"`
-		TrackingId   string `xml:"tracking_id"`
-		Mobile       string `xml:"Mobile"`
-		UDF1         string `xml:"UDF1"`
-		URI          string `xml:"URI"`
-		DocContent   string `xml:"DocContent"`
-		DataContent  string `xml:"DataContent"`
-	} `xml:"DocDetails"`
-}
-
-type PullDocRequest struct {
-	XMLName    xml.Name `xml:"PullDocRequest"`
-	Text       string   `xml:",chardata"`
-	Ns2        string   `xml:"ns2,attr"`
-	Ver        string   `xml:"ver,attr"`
-	Ts         string   `xml:"ts,attr"`
-	Txn        string   `xml:"txn,attr"`
-	OrgId      string   `xml:"orgId,attr"`
-	Format     string   `xml:"format,attr"`
-	DocDetails struct {
-		Text         string `xml:",chardata"`
-		URI          string `xml:"URI"`
-		DigiLockerId string `xml:"DigiLockerId"`
-		UID          string `xml:"UID"`
-		FullName     string `xml:"FullName"`
-		DOB          string `xml:"DOB"`
-	} `xml:"DocDetails"`
-}
-
-type PullDocResponse struct {
-	XMLName        xml.Name `xml:"PullDocResponse"`
-	Text           string   `xml:",chardata"`
-	Ns2            string   `xml:"ns2,attr"`
-	ResponseStatus struct {
-		Text   string `xml:",chardata"`
-		Status string `xml:"Status,attr"`
-		Ts     string `xml:"ts,attr"`
-		Txn    string `xml:"txn,attr"`
-	} `xml:"ResponseStatus"`
-	DocDetails struct {
-		Text        string `xml:",chardata"`
-		DocContent  string `xml:"DocContent"`
-		DataContent string `xml:"DataContent"`
-	} `xml:"DocDetails"`
-}
-
-func ValidMAC(message string, messageMAC [] byte, keyString string) bool {
-	if keyString == "ignore" {
-		return true
-	}
-	key := ([]byte)(keyString)
-	mac := hmac.New(sha256.New, key)
-	mac.Write([]byte(message))
-	expectedMAC := mac.Sum(nil)
-	hexMac := ([]byte)(hex.EncodeToString(expectedMAC))
-	if log.IsLevelEnabled(log.InfoLevel) {
-		log.Infof("Expected mac %s and got %s", base64.StdEncoding.EncodeToString(hexMac), base64.StdEncoding.EncodeToString(messageMAC))
-	}
-	return hmac.Equal(messageMAC, hexMac)
-}
-
-func docRequest(w http.ResponseWriter, req *http.Request) {
-	log.Info("Got document request ")
-	requestBuffer, requestString, hmacSignByteArray, done := preProcessRequest(req, w)
-	if done {
-		return
-	}
-
-	if ValidMAC(requestString, hmacSignByteArray, config.Config.Digilocker.AuthHMACKey) {
-		xmlRequest := PullDocRequest{}
-		if err := xml.Unmarshal(requestBuffer, &xmlRequest); err != nil {
-			log.Errorf("Error in marshalling request from the digilocker %+v", err)
-		} else {
-
-			response := PullDocResponse{}
-			response.ResponseStatus.Ts = xmlRequest.Ts
-			response.ResponseStatus.Txn = xmlRequest.Txn
-			response.ResponseStatus.Status = "0"
-			certBundle := getCertificateByUri(xmlRequest.DocDetails.URI, xmlRequest.DocDetails.DOB)
-			if certBundle != nil {
-				response.ResponseStatus.Status = "1"
-				if xmlRequest.Format == "pdf" || xmlRequest.Format == "both" {
-					if pdfBytes, err := getCertificateAsPdf(certBundle.signedJson); err != nil {
-						log.Errorf("Error in creating certificate pdf %+v", err)
-						go kafkaService.PublishEvent(models.Event{
-							Date:          time.Time{},
-							Source:        "" + xmlRequest.DocDetails.URI,
-							TypeOfMessage: "internal_error",
-							ExtraInfo:     nil,
-						})
-						w.WriteHeader(500)
-					} else {
-						response.DocDetails.DocContent = base64.StdEncoding.EncodeToString(pdfBytes)
-					}
-				}
-				if xmlRequest.Format == "both" || xmlRequest.Format == "xml" {
-					certificateId := certBundle.certificateId
-					xmlCert := "<certificate id=\"" + certificateId + "\"><![CDATA[" + certBundle.signedJson + "]]></certificate>"
-					response.DocDetails.DataContent = base64.StdEncoding.EncodeToString([]byte(xmlCert))
-				}
-
-			} else {
-				go kafkaService.PublishEvent(models.Event{
-					Date:          time.Now(),
-					Source:        xmlRequest.DocDetails.URI,
-					TypeOfMessage: DigilockerFailedEvent,
-					ExtraInfo:     "Certificate not found",
-				})
-			}
-			if responseBytes, err := xml.Marshal(response); err != nil {
-				log.Errorf("Error while serializing xml")
-			} else {
-				go kafkaService.PublishEvent(models.Event{
-					Date:          time.Now(),
-					Source:        xmlRequest.DocDetails.URI,
-					TypeOfMessage: DigilockerSuccessEvent,
-					ExtraInfo:     "Certificate found",
-				})
-				w.Header().Set("Content-Type", "application/xml")
-				w.WriteHeader(200)
-
-				_, _ = w.Write(responseBytes)
-				return
-			}
-			w.WriteHeader(500)
-		}
-	} else {
-		log.Errorf("Unauthorized access")
-		w.WriteHeader(401)
-		_, _ = w.Write([]byte("Unauthorized"))
-		go kafkaService.PublishEvent(models.Event{
-			Date:          time.Now(),
-			Source:        "",
-			TypeOfMessage: DigilockerFailedEvent,
-			ExtraInfo:     "invalid hmac",
-		})
-	}
-}
-
-func uriRequest(w http.ResponseWriter, req *http.Request) {
-	requestBuffer, requestString, hmacSignByteArray, done := preProcessRequest(req, w)
-	if done {
-		return
-	}
-
-	if ValidMAC(requestString, hmacSignByteArray, config.Config.Digilocker.AuthHMACKey) {
-
-		xmlRequest := PullURIRequest{}
-		if err := xml.Unmarshal(requestBuffer, &xmlRequest); err != nil {
-			log.Errorf("Error in marshalling request from the digilocker %+v", err)
-		} else {
-
-			response := PullURIResponse{}
-			response.ResponseStatus.Ts = xmlRequest.Ts
-			response.ResponseStatus.Txn = xmlRequest.Txn
-			response.ResponseStatus.Status = "0"
-			response.DocDetails.DocType = config.Config.Digilocker.DocType
-			response.DocDetails.DigiLockerId = xmlRequest.DocDetails.DigiLockerId
-			response.DocDetails.FullName = xmlRequest.DocDetails.FullName
-			response.DocDetails.DOB = xmlRequest.DocDetails.DOB
-
-			certBundle := getCertificate(xmlRequest.DocDetails.TrackingId, xmlRequest.DocDetails.DOB, xmlRequest.DocDetails.Mobile)
-			if certBundle != nil && certBundle.mobile == xmlRequest.DocDetails.Mobile {
-				response.DocDetails.URI = certBundle.Uri
-				response.ResponseStatus.Status = "1"
-				if xmlRequest.Format == "pdf" || xmlRequest.Format == "both" {
-					if pdfBytes, err := getCertificateAsPdf(certBundle.signedJson); err != nil {
-						log.Errorf("Error in creating certificate pdf")
-					} else {
-						response.DocDetails.DocContent = base64.StdEncoding.EncodeToString(pdfBytes)
-					}
-				}
-				if xmlRequest.Format == "both" || xmlRequest.Format == "xml" {
-					certificateId := certBundle.certificateId
-					xmlCert := "<certificate id=\"" + certificateId + "\"><![CDATA[" + certBundle.signedJson + "]]></certificate>"
-					response.DocDetails.DataContent = base64.StdEncoding.EncodeToString([]byte(xmlCert))
-				}
-
-			} else {
-				go kafkaService.PublishEvent(models.Event{
-					Date:          time.Now(),
-					Source:        xmlRequest.DocDetails.TrackingId,
-					TypeOfMessage: DigilockerFailedEvent,
-					ExtraInfo:     "Certificate not found",
-				})
-			}
-			if responseBytes, err := xml.Marshal(response); err != nil {
-				log.Errorf("Error while serializing xml")
-			} else {
-				go kafkaService.PublishEvent(models.Event{
-					Date:          time.Now(),
-					Source:        xmlRequest.DocDetails.TrackingId,
-					TypeOfMessage: DigilockerSuccessEvent,
-					ExtraInfo:     "Certificate found",
-				})
-				w.Header().Set("Content-Type", "application/xml")
-				w.WriteHeader(200)
-				_, _ = w.Write(responseBytes)
-				return
-			}
-			w.WriteHeader(500)
-		}
-	} else {
-		log.Errorf("Unauthorized access")
-		w.WriteHeader(401)
-		_, _ = w.Write([]byte("Unauthorized"))
-		go kafkaService.PublishEvent(models.Event{
-			Date:          time.Now(),
-			Source:        "",
-			TypeOfMessage: DigilockerFailedEvent,
-			ExtraInfo:     "Invalid hmac",
-		})
-	}
-
-}
-
-func preProcessRequest(req *http.Request, w http.ResponseWriter) ([]byte, string, []byte, bool) {
-	log.Info("Got request ")
-	for name, values := range req.Header {
-		for _, value := range values {
-			log.Infof("%s:%s", name, value)
-		}
-	}
-	requestBuffer := make([]byte, 2048)
-	n, _ := req.Body.Read(requestBuffer)
-	requestString := string(requestBuffer[:n])
-	log.Infof("Read %d bytes ", n)
-	println(requestString)
-	log.Infof("Request body %s", requestString)
-	hmacDigest := req.Header.Get(config.Config.Digilocker.AuthKeyName)
-	hmacSignByteArray, e := base64.StdEncoding.DecodeString(hmacDigest)
-	if e != nil {
-		log.Errorf("Error in verifying request signature")
-		w.WriteHeader(500)
-		_, _ = w.Write([]byte("Error in verifying request signature"))
-		go kafkaService.PublishEvent(models.Event{
-			Date:          time.Now(),
-			Source:        "",
-			TypeOfMessage: DigilockerFailedEvent,
-			ExtraInfo:     "invalid hmac signature",
-		})
-		return nil, "", nil, true
-	}
-	return requestBuffer, requestString, hmacSignByteArray, false
-}
-
-type VaccinationCertificateBundle struct {
-	certificateId string
-	Uri           string
-	signedJson    string
-	mobile        string
-}
-
-func getCertificateByUri(uri string, dob string) *VaccinationCertificateBundle {
-	slice := strings.Split(uri, "-")
-	certificateId := slice[len(slice)-1]
-	certificateFromRegistry, err := getCertificateFromRegistryByCertificateId(certificateId)
-	return returnLatestCertificate(err, certificateFromRegistry, certificateId)
-}
-
-func getCertificate(preEnrollmentCode string, dob string, mobile string) *VaccinationCertificateBundle {
-	certificateFromRegistry, err := getCertificateFromRegistry(preEnrollmentCode)
-	return returnLatestCertificate(err, certificateFromRegistry, preEnrollmentCode)
-}
-
-func returnLatestCertificate(err error, certificateFromRegistry map[string]interface{}, referenceId string) *VaccinationCertificateBundle {
-	if err == nil {
-		certificateArr := certificateFromRegistry[CertificateEntity].([]interface{})
-		if len(certificateArr) > 0 {
-			certificateObj := certificateArr[len(certificateArr)-1].(map[string]interface{})
-			log.Infof("certificate resp %v", certificateObj)
-			var cert VaccinationCertificateBundle
-			cert.certificateId = certificateObj["certificateId"].(string)
-			cert.Uri = "in.gov.covin-" + "VACER" + "-" + cert.certificateId
-			cert.signedJson = certificateObj["certificate"].(string)
-			cert.mobile = certificateObj["mobile"].(string)
-			return &cert
-		} else {
-			log.Errorf("No certificates found for req %v", referenceId)
-		}
-	}
-	return nil
-}
 
 func showLabelsAsPerTemplate(certificate Certificate) []string {
 	if (!isFinal(certificate)) {
@@ -525,20 +189,20 @@ func getCertificateAsPdf(certificateText string) ([]byte, error) {
 	for i = 0; i < rowSize; i++ {
 		pdf.SetX(offsetX)
 		pdf.SetY(offsetY + float64(i)*45.0)
-		pdf.Cell(nil, displayLabels[i])
+		_ = pdf.Cell(nil, displayLabels[i])
 	}
 	pdf.SetX(offsetX)
 	pdf.SetY(offsetY + float64(i)*40.0)
-	pdf.Cell(nil, certificate.CredentialSubject.Address.AddressRegion)
+	_ = pdf.Cell(nil, certificate.CredentialSubject.Address.AddressRegion)
 
 	for i = rowSize; i < len(displayLabels); i++ {
 		pdf.SetX(offsetNewX)
 		pdf.SetY(offsetNewY + float64(i)*45.0)
-		pdf.Cell(nil, displayLabels[i])
+		_ = pdf.Cell(nil, displayLabels[i])
 	}
 	pdf.SetX(offsetNewX)
 	pdf.SetY(offsetNewY + float64(i)*42)
-	pdf.Cell(nil, certificate.Evidence[0].Facility.Address.AddressRegion)
+	_ = pdf.Cell(nil, certificate.Evidence[0].Facility.Address.AddressRegion)
 
 	e := pasteQrCodeOnPage(certificateText, &pdf)
 	if e != nil {
@@ -741,216 +405,6 @@ func getCertificateFromRegistry(preEnrollmentCode string) (map[string]interface{
 	return certificateFromRegistry, err
 }
 
-func getCertificateJSON(w http.ResponseWriter, request *http.Request) {
-	log.Info("GET CERTIFICATE JSON REQUEST")
-	urlParams := request.URL.Query()
-	filter := map[string]interface{}{}
-	preEnrollmentCode := urlParams.Get(PreEnrollmentCode)
-	if preEnrollmentCode != "" {
-		filter[PreEnrollmentCode] = map[string]interface{}{
-			"eq": preEnrollmentCode,
-		}
-	}
-	beneficiaryId := urlParams.Get(BeneficiaryId)
-	if beneficiaryId != "" {
-		filter[BeneficiaryId] = map[string]interface{}{
-			"eq": beneficiaryId,
-		}
-	}
-	if beneficiaryId == "" && preEnrollmentCode == "" {
-		log.Errorf("Get certificates json doesnt contain required fields %v", urlParams.Encode())
-		w.WriteHeader(400)
-		return
-	}
-	certificateFromRegistry, err := services.QueryRegistry(CertificateEntity, filter)
-	if err == nil {
-		certificateArr := certificateFromRegistry[CertificateEntity].([]interface{})
-		log.Infof("Certificate query return %d records", len(certificateArr))
-		if len(certificateArr) > 0 {
-			certificateObj := certificateArr[len(certificateArr)-1].(map[string]interface{})
-			if responseBytes, err := json.Marshal(certificateObj); err != nil {
-				log.Errorf("Error while serializing xml")
-			} else {
-				w.WriteHeader(200)
-				_, _ = w.Write(responseBytes)
-				return
-			}
-		}
-	} else {
-		log.Errorf("No certificates found for request %v", filter)
-	}
-}
-
-func getCertificates(w http.ResponseWriter, request *http.Request) {
-	log.Info("GET CERTIFICATES JSON ")
-	var requestBody map[string]interface{}
-	err := json.NewDecoder(request.Body).Decode(&requestBody)
-	mobile, found := requestBody[Mobile]
-	filter := map[string]interface{}{}
-	beneficiaryId, found := requestBody[BeneficiaryId]
-	if found {
-		filter[PreEnrollmentCode] = map[string]interface{}{
-			"eq": beneficiaryId,
-		}
-	}
-	if mobile == nil || beneficiaryId == nil {
-		log.Errorf("get certificates requested with no parameters, %v", requestBody)
-		w.WriteHeader(400)
-		go kafkaService.PublishEvent(models.Event{
-			Date:          time.Now(),
-			Source:        "",
-			TypeOfMessage: ExternalFailedEvent,
-			ExtraInfo:     "Invalid parameters",
-		})
-		return
-	}
-	certificateFromRegistry, err := services.QueryRegistry(CertificateEntity, filter)
-	if err == nil {
-		certificateArr := certificateFromRegistry[CertificateEntity].([]interface{})
-		log.Infof("Certificate query return %d records", len(certificateArr))
-		if len(certificateArr) > 0 {
-			certificatesForThisMobile := []map[string]interface{}{}
-			for i := 0; i < len(certificateArr); i++ {
-				certificateObj := certificateArr[i].(map[string]interface{})
-				if certificateObj["mobile"] == mobile {
-					certificatesForThisMobile = append(certificatesForThisMobile, certificateObj)
-				}
-			}
-			if len(certificatesForThisMobile) > 0 {
-				certificates := map[string]interface{}{
-					"certificates": certificatesForThisMobile,
-				}
-				if responseBytes, err := json.Marshal(certificates); err != nil {
-					log.Errorf("Error while serializing xml")
-				} else {
-					w.WriteHeader(200)
-					w.Header().Set("Content-Type", "application/json")
-					_, _ = w.Write(responseBytes)
-					go kafkaService.PublishEvent(models.Event{
-						Date:          time.Now(),
-						Source:        pkg.ToString(beneficiaryId),
-						TypeOfMessage: ExternalSuccessEvent,
-						ExtraInfo:     "Certificate found",
-					})
-					return
-				}
-			} else { //no certificate found for this mobile --
-				errorResponse := ErrorResponse{
-					Status:  "not_found",
-					Message: "Mobile number is not matching for the given beneficiary Id",
-				}
-				writeResponse(w, 404, errorResponse)
-			}
-		} else {
-			log.Errorf("No certificates found for request %v", filter)
-			payload := fmt.Sprintf(`No certificate found for the given beneficiary Id %s`, beneficiaryId)
-			errorResponse := ErrorResponse{
-				Status:  "not_found",
-				Message: payload,
-			}
-			writeResponse(w, 404, errorResponse)
-		}
-	} else {
-		log.Errorf("No certificates found for request %v", filter)
-		w.WriteHeader(500)
-	}
-	go kafkaService.PublishEvent(models.Event{
-		Date:          time.Now(),
-		Source:        pkg.ToString(beneficiaryId),
-		TypeOfMessage: ExternalFailedEvent,
-		ExtraInfo:     "Certificate not found",
-	})
-}
-
-func getCertificatePDFHandler(w http.ResponseWriter, r *http.Request) {
-	log.Info("GET PDF HANDLER REQUEST")
-	var requestBody map[string]interface{}
-	err := json.NewDecoder(r.Body).Decode(&requestBody)
-	mobile, found := requestBody[Mobile]
-	filter := map[string]interface{}{}
-	beneficiaryId, found := requestBody[BeneficiaryId]
-	if found {
-		filter[PreEnrollmentCode] = map[string]interface{}{
-			"eq": beneficiaryId,
-		}
-	}
-	if mobile == nil || beneficiaryId == nil {
-		log.Errorf("get certificates requested with no parameters, %v", requestBody)
-		w.WriteHeader(400)
-		go kafkaService.PublishEvent(models.Event{
-			Date:          time.Now(),
-			Source:        "",
-			TypeOfMessage: ExternalFailedEvent,
-			ExtraInfo:     "Invalid parameters",
-		})
-		return
-	}
-	certificateFromRegistry, err := services.QueryRegistry(CertificateEntity, filter)
-	if err == nil {
-		certificateArr := certificateFromRegistry[CertificateEntity].([]interface{})
-		log.Infof("Certificate query return %d records", len(certificateArr))
-		if len(certificateArr) > 0 {
-			certificateObj := certificateArr[len(certificateArr)-1].(map[string]interface{})
-			log.Infof("certificate resp %v", certificateObj)
-			mobileOnCert := certificateObj["mobile"].(string)
-			if mobile != mobileOnCert {
-				statusCode := 404
-				payload := ErrorResponse{
-					Status:  "not_found",
-					Message: `Mobile number is not matching for the given beneficiary Id`,
-				}
-				writeResponse(w, statusCode, payload)
-				return
-			} else {
-				signedJson := certificateObj["certificate"].(string)
-				if pdfBytes, err := getCertificateAsPdf(signedJson); err != nil {
-					log.Errorf("Error in creating certificate pdf")
-				} else {
-					//w.Header().Set("Content-Disposition", "attachment; filename=certificate.pdf")
-					//w.Header().Set("Content-Type", resp.Header.Get("Content-Type"))
-					//w.Header().Set("Content-Length", string(len(pdfBytes)))
-					w.WriteHeader(200)
-					_, _ = w.Write(pdfBytes)
-					go kafkaService.PublishEvent(models.Event{
-						Date:          time.Now(),
-						Source:        pkg.ToString(beneficiaryId),
-						TypeOfMessage: ExternalSuccessEvent,
-						ExtraInfo:     "Certificate found",
-					})
-				}
-			}
-		} else {
-			log.Errorf("No certificates found for request %v", filter)
-			payload := fmt.Sprintf(`No certificate found for the given beneficiary Id %s`, beneficiaryId)
-			errorResponse := ErrorResponse{
-				Status:  "not_found",
-				Message: payload,
-			}
-			writeResponse(w, 404, errorResponse)
-		}
-	} else {
-		log.Infof("Error %+v", err)
-	}
-	go kafkaService.PublishEvent(models.Event{
-		Date:          time.Now(),
-		Source:        pkg.ToString(beneficiaryId),
-		TypeOfMessage: ExternalFailedEvent,
-		ExtraInfo:     "Certificate not found",
-	})
-}
-
-func writeResponse(w http.ResponseWriter, statusCode int, payload ErrorResponse) {
-
-	if payloadBytes, err := json.Marshal(payload); err == nil {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(statusCode)
-		_, _ = w.Write(payloadBytes)
-	} else {
-		log.Errorf("Error in converting response to json %+v", err)
-		w.WriteHeader(500)
-		_, _ = w.Write([]byte("Internal error"))
-	}
-}
 
 func authorize(next http.HandlerFunc, roles []string, eventType string) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
